+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Add sale order and client partner fields on production order.
 - Account Move : Display total number of lines, total debit, total credit and the gap on move form.
 - COST SHEET : cost sheet type
 - Add generated purchase orders panel in sale order form.
+- COST SHEET : calculation date
 
 ## Improvements
 - FINANCIAL ACCOUNT : New boolean 'Analytic Distribution Authorized' added to make Analytic fields required on Invoice Line.